add support for editing diaper events
trying out a potentially bad approach here: combining the new and edit
experiences into one. they really ought to be exactly the same
form-wise. the only differences are cosmetic and the backend save
action. also, you can delete existing events so there's gotta be ui and
behavior for that.

i'm curious to see how this pans out. i think it saves a bunch of
repetitive ui and notifier code.